+++ weblate/formats/base.py
@@ -590,7 +590,7 @@ class TranslationFormat:
         cls,
         base: str,
         monolingual: bool,
-        errors: list | None = None,
+        errors: list[Exception] | None = None,
         fast: bool = False,
     ) -> bool:
         """Check whether base is valid."""