+++ weblate/lang/forms.py
@@ -3,9 +3,12 @@
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 from crispy_forms.helper import FormHelper
+from crispy_forms.layout import Field, Layout
 from django import forms
+from django.conf import settings
 
 from weblate.lang.models import Language, Plural
+from weblate.utils.forms import ContextDiv
 
 
 class LanguageForm(forms.ModelForm):
@@ -17,6 +20,16 @@ class LanguageForm(forms.ModelForm):
         super().__init__(*args, **kwargs)
         self.helper = FormHelper()
         self.helper.form_tag = False
+        self.helper.layout = Layout(
+            ContextDiv(
+                template="lang/language_edit_warning.html",
+                context={"update_languages": settings.UPDATE_LANGUAGES},
+            ),
+            Field("code"),
+            Field("name"),
+            Field("direction"),
+            Field("population"),
+        )
 
     @staticmethod
     def get_field_doc(field):