+++ weblate/trans/models/__init__.py
@@ -20,7 +20,7 @@ from weblate.trans.models.component import Component
 from weblate.trans.models.componentlist import AutoComponentList, ComponentList
 from weblate.trans.models.label import Label
 from weblate.trans.models.pending import PendingUnitChange
-from weblate.trans.models.project import Project
+from weblate.trans.models.project import CommitPolicyChoices, Project
 from weblate.trans.models.suggestion import Suggestion, Vote
 from weblate.trans.models.translation import Translation
 from weblate.trans.models.unit import Unit
@@ -36,6 +36,7 @@ __all__ = [
     "Category",
     "Change",
     "Comment",
+    "CommitPolicyChoices",
     "Component",
     "ComponentList",
     "ContributorAgreement",